Fix too much queue growth
Each call to push added a new node to q.nodes,
regardless of whether there was room in the queue.

In addition, tracking len separately vs. just using a slice type
masked this error, so replace the implementation as a few methods
built around slices. Also, remove the empty slot in the current
implementation.